Separated admin from user
