+++ src/controllers/auth.ts
@@ -53,65 +53,6 @@ export const getMe: RequestHandler = asyncHandler(async (req, res) => {
   });
 });
 
-// @desc    Delete user
-// @route   DELETE /api/v1/auth/users/:id
-// @access  Admin
-export const deleteUser: RequestHandler = asyncHandler(
-  async (req, res, next) => {
-    // This is needed because otherwise you won't be able to get user.role
-    const user = await User.findById(req.params.id);
-
-    // Check if user is trying to delete himself, the check has to be done before you use findByIdAndDelete
-    if (req.params.id === req.user.id) {
-      return next(new ErrorResponse(`You can't delete yourself`, 401));
-    }
-
-    // Check if user exists
-    if (!user) {
-      return next(new ErrorResponse(`User doesn't exist`, 401));
-    }
-
-    // Check if user is another admin
-    if (user.role === 'admin') {
-      return next(new ErrorResponse(`You can not delete other admins`, 401));
-    }
-
-    await User.findByIdAndDelete(req.params.id);
-
-    res.status(200).json({
-      success: true,
-      data: `Deleted user with id of: ${req.params.id}`,
-    });
-  }
-);
-// @desc    Get all users
-// @route   GET /api/v1/auth/users/
-// @access  Admin
-export const getUsers: RequestHandler = asyncHandler(async (req, res) => {
-  const user = await User.find();
-
-  res.status(200).json({
-    success: true,
-    numberOfUsers: user.length,
-    data: user,
-  });
-});
-
-// @desc    Get single user
-// @route   GET /api/v1/auth/users/:id
-// @access  Admin
-export const getUser: RequestHandler = asyncHandler(async (req, res, next) => {
-  const user = await User.findById(req.params.id);
-
-  // If id format is valid but user doesn't exist
-  if (!user) {
-    return next(
-      new ErrorResponse(`User not found with id of ${req.params.id}`, 404)
-    );
-  }
-
-  res.status(200).json({ sucess: true, data: user });
-});
 // @desc    Logged in user edit name and email
 // @route   PUT /api/v1/auth/changedetails
 // @access  Private
@@ -153,18 +94,6 @@ export const updatePassword: RequestHandler = asyncHandler(
     sendTokenResponse(user, 200, res);
   }
 );
-
-// @desc    Edit user
-// @route   PUT /api/v1/users/:id
-// @access  Admin
-export const updateUser: RequestHandler = asyncHandler(async (req, res) => {
-  const user = await User.findByIdAndUpdate(req.params.id, req.body, {
-    new: true,
-    runValidators: true,
-  });
-
-  res.status(200).json({ sucess: true, data: user });
-});
 
 // @desc    Forgot password
 // @route   POST /api/v1/auth/forgotpassword
