+++ src/server.ts
@@ -2,9 +2,10 @@ import express from 'express';
 import dotenv from 'dotenv';
 import { connectDB } from './config/db';
 import cookieParser from 'cookie-parser';
-import productRouter from './routes/product';
 import errorHandler from './middleware/error';
+import productRouter from './routes/product';
 import userRouter from './routes/auth';
+import adminRouter from './routes/admin';
 
 dotenv.config({ path: 'config.env' }); // exporting environment variables
 connectDB(); // connecting to mongoDB
@@ -16,6 +17,7 @@ app.use(cookieParser()); // cookie parser
 
 app.use('/api/v1/products', productRouter);
 app.use('/api/v1/auth', userRouter);
+app.use('/api/v1/admin', adminRouter);
 app.use(errorHandler); // errorHandler has to be after routers
 
 const PORT = process.env.PORT || 3000;
